requirements.txt
Django == 5.1.4
djangorestframework == 3.15.2
psycopg2-binary == 2.9.10
djangorestframework-simplejwt == 5.3.1
django-cors-headers==4.6.0
